Log file functions and the base for the calendar for terminals
An exception template of sorts is provided for consistency. ical_x_defs.hpp will contain all the macros required by libical to have full functionality. The log file functions work as far as I know. 7cterm.cpp is the basis upon which we can build our calendar program. It's basically just a template, but it will contain the ability to add an event (and task?), edit a task, and delete a task. It only keeps content in memory.

Side note: This has been tested with libical-0.44. The latest version on SourceForge is 0.48, but there are some possible errors with writing to files. Compilation is done through g++ 7cterm.cpp -o 7cterm (but the current revision fails to compile due to partially done code.)